Separate all windows function into a separate class
