+++ src/PixiEditor/Helpers/GlobalMouseHook.cs
@@ -15,15 +15,9 @@ public delegate void MouseUpEventHandler(object sender, Point p, MouseButton but
 [ExcludeFromCodeCoverage]
 internal static class GlobalMouseHook
 {
-    private const int WH_MOUSE_LL = 14;
-    private const int WM_LBUTTONUP = 0x0202;
-    private const int WM_MBUTTONUP = 0x0208;
-    private const int WM_RBUTTONUP = 0x0205;
-
     private static int mouseHookHandle;
-    private static HookProc mouseDelegate;
+    private static Win32.HookProc mouseDelegate;
 
-    private delegate int HookProc(int nCode, int wParam, IntPtr lParam);
 
     public static event MouseUpEventHandler OnMouseUp
     {
@@ -56,7 +50,7 @@ internal static class GlobalMouseHook
     {
         if (mouseHookHandle != 0)
         {
-            int result = UnhookWindowsHookEx(mouseHookHandle);
+            int result = Win32.UnhookWindowsHookEx(mouseHookHandle);
             mouseHookHandle = 0;
             mouseDelegate = null;
             if (result == 0)
@@ -72,10 +66,10 @@ internal static class GlobalMouseHook
         if (mouseHookHandle == 0)
         {
             mouseDelegate = MouseHookProc;
-            mouseHookHandle = SetWindowsHookEx(
-                WH_MOUSE_LL,
+            mouseHookHandle = Win32.SetWindowsHookEx(
+                Win32.WH_MOUSE_LL,
                 mouseDelegate,
-                GetModuleHandle(Process.GetCurrentProcess().MainModule.ModuleName),
+                Win32.GetModuleHandle(Process.GetCurrentProcess().MainModule.ModuleName),
                 0);
             if (mouseHookHandle == 0)
             {
@@ -89,60 +83,20 @@ internal static class GlobalMouseHook
     {
         if (nCode >= 0)
         {
-            MSLLHOOKSTRUCT mouseHookStruct = (MSLLHOOKSTRUCT)Marshal.PtrToStructure(lParam, typeof(MSLLHOOKSTRUCT));
-            if (wParam == WM_LBUTTONUP || wParam == WM_MBUTTONUP || wParam == WM_RBUTTONUP)
+            Win32.MSLLHOOKSTRUCT mouseHookStruct = (Win32.MSLLHOOKSTRUCT)Marshal.PtrToStructure(lParam, typeof(Win32.MSLLHOOKSTRUCT));
+            if (wParam == Win32.WM_LBUTTONUP || wParam == Win32.WM_MBUTTONUP || wParam == Win32.WM_RBUTTONUP)
             {
                 if (MouseUp != null)
                 {
 
-                    MouseButton button = wParam == WM_LBUTTONUP ? MouseButton.Left
-                        : wParam == WM_MBUTTONUP ? MouseButton.Middle : MouseButton.Right;
+                    MouseButton button = wParam == Win32.WM_LBUTTONUP ? MouseButton.Left
+                        : wParam == Win32.WM_MBUTTONUP ? MouseButton.Middle : MouseButton.Right;
                     Dispatcher.CurrentDispatcher.BeginInvoke(() =>
                         MouseUp.Invoke(null, new Point(mouseHookStruct.Pt.X, mouseHookStruct.Pt.Y), button));
                 }
             }
         }
 
-        return CallNextHookEx(mouseHookHandle, nCode, wParam, lParam);
-    }
-
-    [DllImport(
-        "user32.dll",
-        CharSet = CharSet.Auto,
-        CallingConvention = CallingConvention.StdCall,
-        SetLastError = true)]
-    private static extern int SetWindowsHookEx(int idHook, HookProc lpfn, IntPtr hMod, int dwThreadId);
-
-    [DllImport(
-        "user32.dll",
-        CharSet = CharSet.Auto,
-        CallingConvention = CallingConvention.StdCall,
-        SetLastError = true)]
-    private static extern int UnhookWindowsHookEx(int idHook);
-
-    [DllImport(
-        "user32.dll",
-        CharSet = CharSet.Auto,
-        CallingConvention = CallingConvention.StdCall)]
-    private static extern int CallNextHookEx(int idHook, int nCode, int wParam, IntPtr lParam);
-
-    [DllImport("kernel32.dll")]
-    private static extern IntPtr GetModuleHandle(string name);
-
-    [StructLayout(LayoutKind.Sequential)]
-    private struct POINT
-    {
-        public int X;
-        public int Y;
-    }
-
-    [StructLayout(LayoutKind.Sequential)]
-    private struct MSLLHOOKSTRUCT
-    {
-        public POINT Pt;
-        public uint MouseData;
-        public uint Flags;
-        public uint Time;
-        public IntPtr DwExtraInfo;
+        return Win32.CallNextHookEx(mouseHookHandle, nCode, wParam, lParam);
     }
 }
